wilfer

+++ src/main/java/com/modelo/prueba.java
@@ -12,6 +12,6 @@ package com.modelo;
 public class prueba {
   
     
-    private int test;
+    private long test;
     
 }
